backend/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.1
PyJWT==2.8.0
werkzeug==3.0.0
python-docx==0.8.11
python-pptx==0.6.21
requests==2.31.0

# IMPORTANT: ADD THIS
google-generativeai==0.4.1
